Prepare to make Next an array of size_t

+++ src/txpost.h
@@ -61,18 +61,18 @@ public:
     [[nodiscard]] auto get_match() const -> Istring { return match; }
 };
 
-// A class to count words...
+// A class to count words... \todo where is that used? \todo chained list???
 class WordList {
     WordList *  next;
     std::string name;
-    int         hash;
+    size_t      hash;
     int         freq;
 
 public:
     void               set_next(WordList *n) { next = n; }
     [[nodiscard]] auto next_empty() const -> bool { return next == nullptr; }
-    WordList(std::string s, int h, WordList *N) : next(N), name(std::move(s)), hash(h), freq(1) {}
-    auto               is_here(String s, int h) const -> bool { return hash == h && strcmp(name.c_str(), s) == 0; }
+    WordList(std::string s, size_t h, WordList *N) : next(N), name(std::move(s)), hash(h), freq(1) {}
+    auto               is_here(String s, size_t h) const -> bool { return hash == h && strcmp(name.c_str(), s) == 0; }
     void               incr_freq() { freq++; }
     [[nodiscard]] auto get_freq() const -> int { return freq; }
     [[nodiscard]] auto get_next() const -> WordList * { return next; }
